Add scripts to automatically update contracts address and abi for frontend project

+++ contracts/ChainLinkPriceFeeds.sol
@@ -19,7 +19,7 @@ library ChainLinkPriceFeeds {
         AggregatorV3Interface priceFeed
     ) internal view returns (uint256) {
         uint256 ethPrice = getPrice(priceFeed);
-        uint ethAmountInUsd = (ethPrice * ethAmount) / 1000000000000000000;
+        uint256 ethAmountInUsd = (ethPrice * ethAmount) / 1000000000000000000;
         return ethAmountInUsd;
     }
 }
